+++ app/cache/cache.py
@@ -3,6 +3,7 @@ from streamlit import session_state as _state
 
 _PERSIST_STATE_KEY = f"{__name__}_PERSIST"
 
+
 def persist(key: str) -> str:
     if _PERSIST_STATE_KEY not in _state:
         _state[_PERSIST_STATE_KEY] = set()
@@ -14,17 +15,21 @@ def persist(key: str) -> str:
 
 def load_widget_state():
     if _PERSIST_STATE_KEY in _state:
-        _state.update({
-            key: value
-            for key, value in _state.items()
-            if key in _state[_PERSIST_STATE_KEY]
-        })
+        _state.update(
+            {
+                key: value
+                for key, value in _state.items()
+                if key in _state[_PERSIST_STATE_KEY]
+            }
+        )
 
-    initCache()
+    init_cache()
 
-def initCache():
-  if "page" not in st.session_state:
-      st.session_state.update({
-        "page": "dashboard",
-    })
 
+def init_cache():
+    if "page" not in st.session_state:
+        st.session_state.update(
+            {
+                "page": "dashboard",
+            }
+        )
