+++ app/service/delivery_api.py
@@ -1,7 +1,9 @@
 from requests import get
 
-def getAllDeliveryMan():
-    return get('https://cybersecurity-fiap.herokuapp.com/v1/deliveryman').json()
 
-def getAllBuyers():
-    return get('https://cybersecurity-fiap.herokuapp.com/v1/buyer').json()
+def get_all_deliveryman():
+    return get("https://cybersecurity-fiap.herokuapp.com/v1/deliveryman").json()
+
+
+def get_all_buyers():
+    return get("https://cybersecurity-fiap.herokuapp.com/v1/buyer").json()
