+++ app/view/components.py
@@ -1,0 +1,5 @@
+from app.view.dashboard import dashboard
+
+PAGES = {
+    "dashboard": dashboard,
+}
