--- app/view/components/components.py
@@ -1,5 +1,0 @@
-from app.view.components.dashboard.dashboard import dashboard
-
-PAGES = {
-      "dashboard": dashboard,
-}
