+++ run.py
@@ -1,4 +1,4 @@
-from app.view.components.main import create
+from app.main import create
 
 if __name__ == '__main__':
     create()
